renaming dc_configure to configure and renaming the autoconfigure modules
