+++ src/dc_imex.rs
@@ -10,7 +10,7 @@ use crate::chat;
 use crate::config::Config;
 use crate::constants::*;
 use crate::context::Context;
-use crate::dc_configure::*;
+use crate::configure::*;
 use crate::dc_e2ee::*;
 use crate::dc_msg::*;
 use crate::dc_tools::*;
