+++ src/job.rs
@@ -8,7 +8,7 @@ use rand::{thread_rng, Rng};
 use crate::chat;
 use crate::constants::*;
 use crate::context::Context;
-use crate::dc_configure::*;
+use crate::configure::*;
 use crate::dc_imex::*;
 use crate::dc_location::*;
 use crate::dc_loginparam::*;
